Extend implementation

+++ OBPostupyApi/OBPostupyApi/Repositories/RaceRepository.cs
@@ -5,6 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Dapper;
 
 namespace OBPostupyApi.Repositories
 {
@@ -55,16 +56,19 @@ namespace OBPostupyApi.Repositories
                 .Include(r => r.Categories)
                 .Include(r => r.Maps)
                 .SingleOrDefaultAsync();
+            
+            if (race?.CourseData != null)
+            {
+                var controls = await _context.Controls.Where(c => c.CourseDataId == race.CourseData.Id).ToListAsync();
+                var courses = await _context.Courses.Where(c => c.CourseDataId == race.CourseData.Id).ToListAsync();
+                var splits = await _context.Splits.Where(c => c.CourseDataId == race.CourseData.Id).ToListAsync();
 
-            var controls = await _context.Controls.Where(c => c.CourseDataId == race.CourseData.Id).ToListAsync();
-            var courses = await _context.Courses.Where(c => c.CourseDataId == race.CourseData.Id).ToListAsync();
-            var splits = await _context.Splits.Where(c => c.CourseDataId == race.CourseData.Id).ToListAsync();
-
-            _context.Controls.RemoveRange(controls);
-            _context.Courses.RemoveRange(courses);
-            _context.Splits.RemoveRange(splits);
-            _context.Categories.RemoveRange(race.Categories);
-            _context.Maps.RemoveRange(race.Maps);
+                _context.Controls.RemoveRange(controls);
+                _context.Courses.RemoveRange(courses);
+                _context.Splits.RemoveRange(splits);
+            }
+            if (race?.Categories != null) _context.Categories.RemoveRange(race.Categories);
+            if (race?.Maps != null) _context.Maps.RemoveRange(race.Maps);
             _context.Races.Remove(race);
 
             await SaveAsync();
@@ -82,6 +86,25 @@ namespace OBPostupyApi.Repositories
         public async Task SaveAsync()
         {
             await _context.SaveChangesAsync();
+        }
+
+        public async Task<List<Race>> GetAllUserRacesByRegNumberAsync(string regNumber)
+        {
+            return await _context.Person
+                .FromSqlRaw($"SELECT * FROM Person WHERE RegNumbers = '{regNumber}'")
+                .Include(p => p.PersonResults)
+                .ThenInclude(pp => pp.Category)
+                .ThenInclude(c => c.Race)
+                .SelectMany(p => p.PersonResults.Select(pp => pp.Category).Select(c => c.Race))
+                .Where(r => r.Type == RaceType.Public)
+                .ToListAsync();
+        }
+
+        public async Task<List<Race>> GetAllRacesAsync()
+        {
+            return await _context.Races
+                .Include(r => r.User)
+                .ToListAsync();
         }
     }
 }
